Add CreateAssignmentContainer and ListMyAssignment components
