+++ src/components/Lab/Assignment/CreateAssignment/CreateAssignmentContainer/index.tsx
@@ -2,35 +2,55 @@ import { useState } from "react";
 import { useQueryClient } from "react-query";
 import { toast } from "react-toastify";
 import { BulkAssignment, Lab, Profile } from "../../../../../dataStructures";
-import { useCreateAssignments } from "../../../../../hooks/useAssignment";
+import { useCreateAssignments, useCreateMyAssignments } from "../../../../../hooks/useAssignment";
+import { useGetMyProfile } from "../../../../../hooks/useProfile";
 import Button from "../../../../UserInterfaceComponents/Button";
 import Container from "../../../../UserInterfaceComponents/Container";
 import SelectLabsDropdown from "../SelectLabsDropdown";
 import SelectProfilesDropdown from "../SelectProfilesDropdown";
 
-export default function CreateAssignmentContainer() {
+type Props = {
+  collapsed?: boolean;
+};
+
+export default function CreateAssignmentContainer({ collapsed = false }: Props) {
   const { mutateAsync: createBulkAssignments } = useCreateAssignments();
+  const { mutateAsync: createMyBulkAssignments } = useCreateMyAssignments();
+  const { data: myProfile } = useGetMyProfile();
+
   const queryClient = useQueryClient();
 
   const [selectedLabs, setSelectedLabs] = useState<Lab[]>([]);
   const [selectedProfiles, setSelectedProfiles] = useState<Profile[]>([]);
 
   function onAssignClick() {
+    if (myProfile === undefined) {
+      toast.error("not able to pull profile to assign");
+      return;
+    }
+
     let bulkAssignments: BulkAssignment = {
       labIds: selectedLabs.map((lab) => lab.id),
       userIds: selectedProfiles.map((profile) => profile.userPrincipal),
     };
 
-    const response = toast.promise(createBulkAssignments(bulkAssignments), {
-      pending: "Creating assignments...",
-      success: "Assignments created!",
-      error: {
-        render(data: any) {
-          return `Error creating assignments. ${data.data.response.data.error}`;
+    // let response: Promise<AxiosResponse<any, any>>;
+
+    const response = toast.promise(
+      myProfile.roles.includes("mentor")
+        ? createBulkAssignments(bulkAssignments)
+        : createMyBulkAssignments(bulkAssignments),
+      {
+        pending: "Creating assignments...",
+        success: "Assignments created!",
+        error: {
+          render(data: any) {
+            return `Error creating assignments. ${data.data.response.data.error}`;
+          },
+          autoClose: 5000,
         },
-        autoClose: 5000,
-      },
-    });
+      }
+    );
     response.finally(() => {
       // invalidate all lab id queries.
       selectedLabs.forEach((lab) => {
@@ -41,14 +61,8 @@ export default function CreateAssignmentContainer() {
       selectedProfiles.forEach((user) => {
         queryClient.invalidateQueries(["get-assignments-by-user-id", user]);
         queryClient.invalidateQueries(["get-assignments-by-user-id", user]);
-        queryClient.invalidateQueries([
-          "get-readiness-labs-redacted-by-user-id",
-          user,
-        ]);
-        queryClient.invalidateQueries([
-          "get-readiness-labs-redacted-by-user-id",
-          "my",
-        ]);
+        queryClient.invalidateQueries(["get-readiness-labs-redacted-by-user-id", user]);
+        queryClient.invalidateQueries(["get-readiness-labs-redacted-by-user-id", "my"]);
         queryClient.invalidateQueries("get-my-assignments");
       });
 
@@ -58,16 +72,10 @@ export default function CreateAssignmentContainer() {
   }
 
   return (
-    <Container title="Create Assignment">
+    <Container title="Create Assignment" collapsible={true} collapsed={collapsed}>
       <div className="mb-4 flex w-full flex-col justify-between gap-4 bg-slate-50 dark:bg-slate-900 md:flex-row">
-        <SelectLabsDropdown
-          selectedLabs={selectedLabs}
-          setSelectedLabs={setSelectedLabs}
-        />
-        <SelectProfilesDropdown
-          selectedProfiles={selectedProfiles}
-          setSelectedProfiles={setSelectedProfiles}
-        />
+        <SelectLabsDropdown selectedLabs={selectedLabs} setSelectedLabs={setSelectedLabs} />
+        <SelectProfilesDropdown selectedProfiles={selectedProfiles} setSelectedProfiles={setSelectedProfiles} />
         <div className="flex">
           <Button variant="primary-outline" onClick={onAssignClick}>
             Assign
