+++ src/pages/LabsGridPage/index.tsx
@@ -1,6 +1,8 @@
 import { useEffect, useState } from "react";
 import { useParams } from "react-router-dom";
 import SelectedDeployment from "../../components/Deployments/SelectedDeployment";
+import CreateAssignmentContainer from "../../components/Lab/Assignment/CreateAssignment/CreateAssignmentContainer";
+import ListMyAssignment from "../../components/Lab/Assignment/ListMyAssignment";
 import LabCard from "../../components/Lab/LabCard";
 import Terminal from "../../components/Terminal";
 import FilterTextBox from "../../components/UserInterfaceComponents/FilterTextBox";
@@ -23,13 +25,9 @@ export default function LabsGridPage() {
 
   useEffect(() => {
     if (type.endsWith("lab")) {
-      setPageHeading(
-        type.charAt(0).toUpperCase() + type.slice(1).replace(/lab$/, " Labs")
-      );
+      setPageHeading(type.charAt(0).toUpperCase() + type.slice(1).replace(/lab$/, " Labs"));
     } else if (type.endsWith("case")) {
-      setPageHeading(
-        type.charAt(0).toUpperCase() + type.slice(1).replace(/case$/, " Cases")
-      );
+      setPageHeading(type.charAt(0).toUpperCase() + type.slice(1).replace(/case$/, " Cases"));
     } else if (type.endsWith("assignment")) {
       setPageHeading("Assignments");
     } else if (type.endsWith("challenge")) {
@@ -44,9 +42,7 @@ export default function LabsGridPage() {
   const handleSearchChange = (value: string) => setSearchTerm(value);
 
   const filteredLabs = labs?.filter((lab) =>
-    Object.values(lab).some((value) =>
-      value?.toString().toLowerCase().includes(searchTerm.toLowerCase())
-    )
+    Object.values(lab).some((value) => value?.toString().toLowerCase().includes(searchTerm.toLowerCase()))
   );
 
   if (isLoading)
@@ -56,7 +52,7 @@ export default function LabsGridPage() {
       </PageLayout>
     );
 
-  if (!labs?.length) {
+  if (!labs?.length && !type.endsWith("assignment")) {
     return (
       <PageLayout heading={pageHeading}>
         <p className="text-4xl">No {pageHeading.toLowerCase()} found!</p>
@@ -67,6 +63,14 @@ export default function LabsGridPage() {
   return (
     <PageLayout heading={pageHeading}>
       <SelectedDeployment sticky={false} />
+      {type.endsWith("assignment") && (
+        <>
+          <CreateAssignmentContainer collapsed={labs && labs?.length > 0} />
+          <div className="mb-3"></div>
+          <ListMyAssignment />
+          <div className="mb-3"></div>
+        </>
+      )}
       <FilterTextBox
         value={searchTerm}
         onChange={(value: string) => handleSearchChange(value)}
